Exercise 6 custom useEscapeKey hook
