+++ src/components/ToastProvider/ToastProvider.js
@@ -1,23 +1,15 @@
 import React from 'react';
+import useEscapeKey from '../../hooks/use-escape-key';
 
 export const ToastContext = React.createContext();
 
 function ToastProvider({ children }) {
 	const [toasts, setToasts] = React.useState([]);
+	function onEscape() {
+		setToasts([]);
+	}
 
-	React.useEffect(() => {
-		function handleKeyDown(event) {
-			if (event.code === 'Escape') {
-				setToasts([]);
-			}
-		}
-
-		window.addEventListener('keydown', handleKeyDown);
-
-		return () => {
-			window.removeEventListener('keydown', handleKeyDown);
-		};
-	}, []);
+	useEscapeKey(onEscape);
 
 	function createToast(message, variant) {
 		const newToast = {
